+++ include/MHMemory.hpp
@@ -1,11 +1,11 @@
 #pragma once
-#include <experimental/filesystem>
+#include "filesystem.hpp"
 
 #include "PlayerData.hpp"
 
-namespace fs = std::experimental::filesystem;
+namespace fs = filesystem;
 
-fs::path CurrentExecutableDir();
+fs::Path CurrentExecutableDir();
 
 class MH_Memory
 {
@@ -17,29 +17,31 @@ public:
     bool DataAddressFound() const { return _DataPtr != 0; }
 
     int getSteamID() const { return _SteamID;}
-    fs::path getSteamPath() const { return _SteamPath; }
+    fs::Path getSteamPath() const { return _SteamPath; }
     PlayerData getPlayerData() const { return _Data; }
     PlayerData& getPlayerData() { return _Data; }
     Process getProcess() const { return _MHProcess; }
 
-    void setSteamDirectory(const fs::path &Path);
+    void setSteamDirectory(const fs::Path &Path);
 
     bool FetchPlayerData(int slot);
     void FindAddress(std::string Ver="Latest");
     bool BackupSaveData() const ;
     bool WriteArmor(int CharSlot, bool isSafe = true);
 
-    fs::path ExeFilePath = CurrentExecutableDir();
-    fs::path BACKUP_DIR = CurrentExecutableDir().append("Backups");
-    fs::path LogPath = CurrentExecutableDir().append("CostumeArmor.log");
+    fs::Path ExeFilePath = CurrentExecutableDir();
+    fs::Path BACKUP_DIR = CurrentExecutableDir().append("Backups");
+    fs::Path LogPath = CurrentExecutableDir().append("CostumeArmor.log");
 
-    static std::map<std::string,int> Versions;
+    static std::map<std::string, SearchPattern> Versions;
+    static std::map<std::string, int> CharSlotDist;
 
   private:
     Process _MHProcess;
     int _SteamID = 0;
-    fs::path _SteamPath;
+    fs::Path _SteamPath;
     PlayerData _Data;
     DWORD64 _DataPtr = 0;
     bool _SteamFound = false;
+    int _slotDist;
 };