+++ src/MHMemory.cpp
@@ -8,11 +8,12 @@ INITIALIZE_EASYLOGGINGPP
 /// Begin MH_Memory Member definitions
 MH_Memory::MH_Memory(const std::string &ProcName, const std::string &SteamDLL) : _MHProcess(ProcName)
 {
-    INIT_LOGGER( LogPath.string() );
+    INIT_LOGGER( LogPath.str() );
 
     if (!_MHProcess.isOpen())
         return;
 
+    this->_slotDist = MH_Memory::CharSlotDist["Latest"];
     auto Mod = _MHProcess.getModuleByName(SteamDLL);
     if (!Mod.isEmpty())
     {
@@ -20,9 +21,6 @@ MH_Memory::MH_Memory(const std::string &ProcName, const std::string &SteamDLL) :
         _SteamPath = GetRegKeyValue(HKEY_CURRENT_USER, "Software\\Valve\\Steam", "SteamPath");
         if (_SteamID != 0 && !_SteamPath.empty() )
         {
-            #ifdef NDEBUG
-            _SteamPath.make_preferred();
-            #endif
             _SteamPath /= "userdata";
             _SteamPath /= std::to_string(_SteamID);
             _SteamPath /= "582010";
@@ -49,7 +47,7 @@ MH_Memory::MH_Memory(const std::string &ProcName, const std::string &SteamDLL) :
 
 void MH_Memory::FindAddress(std::string Ver)
 {
-    if ( !this->ProcessOpen() )
+    if (!this->ProcessOpen())
         return;
     
     if (MH_Memory::Versions.find(Ver) == MH_Memory::Versions.end()) // Safety Check
@@ -59,6 +57,7 @@ void MH_Memory::FindAddress(std::string Ver)
     if (_DataPtr == 0)
         return;
     _DataPtr -= 29;
+    _slotDist = MH_Memory::CharSlotDist[Ver];
 
     DEBUG_LOG_HEX(DEBUG,"Address of the Characters Data : "<< _DataPtr);
 }
@@ -72,9 +71,9 @@ bool MH_Memory::FetchPlayerData(int slot)
         slot = 0;
     if (slot > 2)
         slot = 2;
-    
-    byte *CharDataBuffer = _MHProcess.ReadMemory(_DataPtr + 1285888 * slot, 28);
-    byte *lpBuffer = _MHProcess.ReadMemory(_DataPtr + 1285888 * slot - 394460, 1);
+
+    byte *CharDataBuffer = _MHProcess.ReadMemory(_DataPtr + _slotDist * slot, 28);
+    byte *lpBuffer = _MHProcess.ReadMemory(_DataPtr + _slotDist * slot - 394460, 1);
 
     if (!CharDataBuffer || !lpBuffer)
     {
@@ -89,12 +88,9 @@ bool MH_Memory::FetchPlayerData(int slot)
     return true;
 }
 
-void MH_Memory::setSteamDirectory(const fs::path &Path)
+void MH_Memory::setSteamDirectory(const fs::Path &p)
 {
-    this->_SteamPath = Path;
-    #ifdef NDEBUG
-    this->_SteamPath.make_preferred();
-    #endif
+    this->_SteamPath = p;
     _SteamPath /= "userdata";
     _SteamPath /= std::to_string(_SteamID);
     _SteamPath /= "582010";
@@ -121,7 +117,7 @@ bool MH_Memory::BackupSaveData() const
         return false;
     }
 
-    if (!fs::exists(MH_Memory::BACKUP_DIR))
+    if (!MH_Memory::BACKUP_DIR.exists())
         if (!fs::create_directory(MH_Memory::BACKUP_DIR))
         {
             DEBUG_LOG(ERROR,"Backup Dir was not Found and couldn't created");
@@ -132,7 +128,7 @@ bool MH_Memory::BackupSaveData() const
     SourcePath /= "remote";
     SourcePath /= "SAVEDATA1000";
 
-    if (!fs::exists(SourcePath))
+    if (!SourcePath.exists())
     {
         DEBUG_LOG(ERROR,"Couldn't Find Save Data");
         return false;
@@ -142,15 +138,15 @@ bool MH_Memory::BackupSaveData() const
     DestPath /= ("Backup " + GetDateTime());
     try
     {
-        fs::copy(SourcePath, DestPath);
+        fs::copy_file(SourcePath, DestPath);
     }
-    catch (fs::filesystem_error &e)
+    catch (fs::FilesystemError &e)
     {
         DEBUG_LOG(ERROR,"Couldn't Copy Save data. Error : " << e.what() );
         return false;
     }
 
-    return fs::exists(DestPath);
+    return DestPath.exists();
 }
 
 bool MH_Memory::WriteArmor(int CharSlot, bool isSafe)
@@ -176,42 +172,48 @@ bool MH_Memory::WriteArmor(int CharSlot, bool isSafe)
 
     for (int index = 0; index < 5; ++index)
     {
-        Status &= _MHProcess.WriteMemoryUInt(_DataPtr + index * 4 + 1285888 * CharSlot, _Data.getArmorPiece(index));
+        Status &= _MHProcess.WriteMemoryUInt(_DataPtr + index * 4 + _slotDist * CharSlot,
+                                            _Data.getArmorPiece(index));
     }
     
     std::stringstream S;
     for( const auto d : _Data.getData() )
         S << ( (int)d ) << ", ";
-    DEBUG_LOG(DEBUG,"Writed (" << S.str()<<") to the game." );
+    DEBUG_LOG(DEBUG,"Written (" << S.str()<<") to the game." );
 
     return Status;
 }
 
-// ByteArray to search
-// Version : 163956, Byte Arr: [231,188,66,1]
-// Version : 165889, Byte Arr: [174,190,66,1]
-// Version : 166849, Byte Arr: [ 47,192,66,1]
-// Version : 167353, Byte Arr: [ 55,193,66,1]
+// ByteArray to search Given version
+// Version : 163956, Byte Arr: [231,188,66,1], last bits: (0x068C) + 29
+// Version : 165889, Byte Arr: [174,190,66,1], last bits: (0x068C) + 29
+// Version : 166849, Byte Arr: [ 47,192,66,1], last bits: (0x068C) + 29
+// Version : 167353, Byte Arr: [ 55,193,66,1], last bits: (0x06AC) + 29
+std::map<std::string, SearchPattern> MH_Memory::Versions{
+    {"163956", {BytesToInt({231, 188, 66, 1}), (0x068C)+29}},
+    {"165889", {BytesToInt({174, 190, 66, 1}), (0x068C)+29}},
+    {"166849", {BytesToInt({ 47, 192, 66, 1}), (0x068C)+29}},
+    {"167353", {BytesToInt({ 55, 193, 66, 1}), (0x068C)+29}},
+    {"167541", {BytesToInt({214, 199, 66, 1}), (0x06AC)+29}},
+    {"Latest", {BytesToInt({214, 199, 66, 1}), (0x06AC)+29}}
+};
 
-// Byte pattern to search given version
-std::map<std::string, int> MH_Memory::Versions{
-    {"163956", BytesToInt({231, 188, 66, 1})},
-    {"165889", BytesToInt({174, 190, 66, 1})},
-    {"166849", BytesToInt({ 47, 192, 66, 1})},
-    {"167353", BytesToInt({ 55, 193, 66, 1})},
-    {"167541", BytesToInt({ 214, 199, 66, 1})},
-    {"Latest", BytesToInt({ 214, 199, 66, 1})}};
+// The distance between the address of the data of the character slots
+std::map<std::string, int> MH_Memory::CharSlotDist{
+    {"163956", 1285888},
+    {"165889", 1285888},
+    {"166849", 1285888},
+    {"167353", 1285888},
+    {"167541", 1285920},
+    {"Latest", 1285920}};
 
 /// End MH_Memory Member definitions
-fs::path CurrentExecutableDir()
+fs::Path CurrentExecutableDir()
 {
-    TCHAR Buffer[1024];
-    if (GetModuleFileName(nullptr, Buffer, 1024) != 0)
+    wchar_t Buffer[1024];
+    if (GetModuleFileNameW(nullptr, Buffer, 1024) != 0)
     {
-        fs::path Path(Buffer);
-        #ifdef NDEBUG
-        Path.make_preferred();
-        #endif
+        fs::Path Path(Buffer);
         return Path.parent_path();
     }
     DEBUG_LOG(ERROR, "Couldn't retrieve the name of current EXE File.");