+++ src/PlayerData.cpp
@@ -64,14 +64,11 @@ std::ostream &operator<<(std::ostream &out, PlayerData &Play)
 
 // Most important function.
 // Searches the memory for a pattern (that changes each version)
-DWORD64 FindDataAddress(Process &Proc, int IntPattern)
+DWORD64 FindDataAddress(Process &Proc, SearchPattern Pa)
 {
     byte PatternBuffer[4];
-    int LastBits;
-    if (IntPattern >= BytesToInt({214, 199, 66, 1}))
-        LastBits = (0x06AC) + 29; // The value of the least significant bits of the address
-    else
-        LastBits = (0x068C) + 29; // The value of the least significant bits of the address
+    int LastBits = Pa.last_bits;
+    int IntPattern = Pa.int_pattern;
 
     DWORD64 BaseAddr = 0;
     byte *ReadBuffer = nullptr;